Remove `CompositeMutationGuard`
