+++ subprojects/core/src/main/java/org/gradle/api/internal/CompositeDomainObjectSet.java
@@ -37,7 +37,7 @@ import java.util.Set;
  *
  * @param <T> The type of domain objects in the component collections of this collection.
  */
-public class CompositeDomainObjectSet<T> extends DelegatingDomainObjectSet<T> implements WithEstimatedSize, WithMutationGuard {
+public class CompositeDomainObjectSet<T> extends DelegatingDomainObjectSet<T> implements WithEstimatedSize {
 
     private final Spec<T> uniqueSpec = new ItemIsUniqueInCompositeSpec();
     private final Spec<T> notInSpec = new ItemNotInCompositeSpec();
@@ -309,7 +309,7 @@ public class CompositeDomainObjectSet<T> extends DelegatingDomainObjectSet<T> im
 
         @Override
         public MutationGuard getMutationGuard() {
-            return MutationGuards.of(store);
+            throw new UnsupportedOperationException();
         }
     }
 }
